// frontend/src/api/generated/models/productionOrderResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface ProductionOrderResponse {
  id: string;
  order_number: string;
  order_type: string;
  sales_po_line_id?: string | null;
  product_variant_id: string;
  product_variant_sku?: string | null;
  target_quantity: number;
  status: string;
  target_completion_date?: string | null;
}